# openapi/wsapi-api.yml
openapi: 3.0.3
info:
  title: WSAPI - WhatsApp API
  description: RESTful HTTP API for WhatsApp via whatsmeow. Supports multi-instance connections, webhook event delivery, and HMAC-SHA256 signing.
  version: 2.0.0
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8080
    description: Local development server
...
security:
  - InstanceAuth: []
paths:
...
  /session/qr:
    get:
      tags: [Session]
      summary: Get QR code for pairing
      description: Returns a PNG image of the QR code to scan with WhatsApp.
      operationId: getQR
      parameters:
        - $ref: "#/components/parameters/XInstanceId"
      responses:
        "200":
          description: QR code PNG image
          content:
            image/png:
              schema:
                type: string
                format: binary
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
  /session/qr/text:
    get:
      tags: [Session]
      summary: Get QR code as text
      description: Returns the raw QR code string for clients that want to render the QR code themselves.
      operationId: getQRText
      parameters:
        - $ref: "#/components/parameters/XInstanceId"
      responses:
        "200":
          description: QR code string
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: string
                    description: Raw QR code string
                required:
                  - code
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
  /session/pair-code/{phone}:
    get:
      tags: [Session]
      summary: Get pair code for phone number
      operationId: getPairCode
      parameters:
        - $ref: "#/components/parameters/XInstanceId"
        - name: phone
          in: path
          required: true
          description: Phone number to pair with (e.g. 1234567890)
          schema:
            type: string
            pattern: '^\+?\d{7,15}$'
      responses:
        "200":
          description: Pair code generated
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: string
                    example: "ABCD-EFGH"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
  /session/status:
    get:
      tags: [Session]
      summary: Get session status
      operationId: getSessionStatus
      parameters:
        - $ref: "#/components/parameters/XInstanceId"
      responses:
        "200":
          description: Session status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SessionStatus"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
  /session/logout:
    post:
      tags: [Session]
      summary: Logout from WhatsApp
      operationId: logout
      parameters:
        - $ref: "#/components/parameters/XInstanceId"
      responses:
        "204":
          description: Logged out
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
  /session/flush-history:
    post:
      tags: [Session]
      summary: Flush cached history sync messages
      description: |
        Returns 202 Accepted immediately, then asynchronously publishes cached history sync
        messages as `message_history_sync` events (one per chat, chunked at 500 messages).
        The consumer should call this endpoint after receiving the `initial_sync_finished` event,
        which is only emitted when `historySync` is enabled and fires after the RECENT history
        sync completes and all messages have been cached.
        Requires the instance to have `historySync` enabled and the device to be paired.
        Cached messages expire after 1 hour.
      operationId: flushHistory
      parameters:
        - $ref: "#/components/parameters/XInstanceId"
      responses:
        "202":
          description: History flush initiated
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: ok
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"

  # ─── Messages ──────────────────────────────────────────────────────
...
  /newsletters:
    get:
      tags: [Newsletters]
      summary: List subscribed newsletters
      description: Returns all newsletters the current user is subscribed to.
      operationId: listNewsletters
      parameters:
        - $ref: "#/components/parameters/XInstanceId"
      responses:
        "200":
          description: List of newsletters
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/NewsletterInfoResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "409":
          $ref: "#/components/responses/DeviceNotPaired"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"

    post:
      tags: [Newsletters]
      summary: Create a newsletter
      description: Create a new newsletter/channel with the specified name, description, and optional picture.
      operationId: createNewsletter
      parameters:
        - $ref: "#/components/parameters/XInstanceId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateNewsletterRequest"
            example:
              name: "My Newsletter"
              description: "Updates about cool stuff"
      responses:
        "201":
          $ref: "#/components/responses/Created"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "409":
          $ref: "#/components/responses/DeviceNotPaired"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
  /newsletters/invite/{code}:
    get:
      tags: [Newsletters]
      summary: Get newsletter info by invite code
      description: Look up newsletter information using an invite code.
      operationId: getNewsletterByInviteCode
      parameters:
        - $ref: "#/components/parameters/XInstanceId"
        - name: code
          in: path
          required: true
          description: Newsletter invite code
          schema:
            type: string
      responses:
        "200":
          description: Newsletter info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NewsletterInfoResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/DeviceNotPaired"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
  /newsletters/{id}:
    get:
      tags: [Newsletters]
      summary: Get newsletter info
      description: Retrieve detailed information about a specific newsletter.
      operationId: getNewsletter
      parameters:
        - $ref: "#/components/parameters/XInstanceId"
        - $ref: "#/components/parameters/NewsletterIdPath"
      responses:
        "200":
          description: Newsletter info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NewsletterInfoResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/DeviceNotPaired"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
  /newsletters/{id}/subscription:
    put:
      tags: [Newsletters]
      summary: Subscribe or unsubscribe from a newsletter
      description: Set the subscription state for a newsletter. Set `subscribed` to true to follow/subscribe, or false to unfollow/unsubscribe.
      operationId: setNewsletterSubscription
      parameters:
        - $ref: "#/components/parameters/XInstanceId"
        - $ref: "#/components/parameters/NewsletterIdPath"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SetSubscriptionRequest"
            examples:
              subscribe:
                summary: Subscribe to newsletter
                value:
                  subscribed: true
              unsubscribe:
                summary: Unsubscribe from newsletter
                value:
                  subscribed: false
      responses:
        "204":
          description: Subscription state updated
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "409":
          $ref: "#/components/responses/DeviceNotPaired"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
  /newsletters/{id}/mute:
    put:
      tags: [Newsletters]
      summary: Mute or unmute a newsletter
      description: Toggle mute state for a newsletter. Muted newsletters do not send notifications.
      operationId: toggleMuteNewsletter
      parameters:
        - $ref: "#/components/parameters/XInstanceId"
        - $ref: "#/components/parameters/NewsletterIdPath"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ToggleMuteNewsletterRequest"
            examples:
              mute:
                summary: Mute newsletter
                value:
                  mute: true
              unmute:
                summary: Unmute newsletter
                value:
                  mute: false
      responses:
        "204":
          description: Mute state updated
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "409":
          $ref: "#/components/responses/DeviceNotPaired"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"

  # ─── Status ──────────────────────────────────────────────────────
...
components:
  securitySchemes:
    AdminAuth:
      type: apiKey
      in: header
      name: X-Api-Key
      description: Admin API key for instance management endpoints (`/admin/instances/*`). Set via `auth.adminApiKey` config.

    InstanceAuth:
      type: apiKey
      in: header
      name: X-Api-Key
      description: Instance API key. Resolved from the instance's own `apiKey` or the global `auth.defaultApiKey`.
  parameters:
    XInstanceId:
      name: X-Instance-Id
      in: header
      required: true
      description: The ID of the WhatsApp instance to use
      schema:
        type: string
        default: "{{instanceId}}"
...
    NewsletterIdPath:
      name: id
      in: path
      required: true
      description: Newsletter JID
      schema:
        type: string
...
  responses:
    Created:
      description: Resource created
      content:
        application/json:
          schema:
            type: object
            properties:
              id:
                type: string
                description: ID of the created resource
          example:
            id: "01234567890123456"
...
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            status: 400
            detail: "invalid request body"
    Unauthorized:
      description: Authentication failed
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            status: 401
            detail: "unauthorized"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            status: 404
            detail: "not found"
    ServiceUnavailable:
      description: Instance not available (not connected or service is nil)
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            status: 503
            detail: "instance not available"
    DeviceNotPaired:
      description: Device is not paired. Complete the pairing process first.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            status: 403
            detail: "device not paired"
  schemas:
    Error:
      type: object
      description: Standard error response returned by all endpoints on failure.
      required: [status, detail]
      properties:
        status:
          type: integer
          description: HTTP status code
          example: 400
        detail:
          type: string
          description: Human-readable error message
          example: "invalid request body"

    # ─── Instance Schemas ──────────────────────────────────────────
...
    SessionStatus:
      type: object
      properties:
        isConnected:
          type: boolean
        isLoggedIn:
          type: boolean
        deviceId:
          type: string

    # ─── Message Schemas ───────────────────────────────────────────
...
    CreateNewsletterRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
          description: Newsletter name
        description:
          type: string
          description: Newsletter description
        picture:
          type: string
          description: Base64-encoded profile picture for the newsletter
    SetSubscriptionRequest:
      type: object
      required: [subscribed]
      properties:
        subscribed:
          type: boolean
          description: True to subscribe, false to unsubscribe
    ToggleMuteNewsletterRequest:
      type: object
      properties:
        mute:
          type: boolean
          description: True to mute, false to unmute
    NewsletterInfoResponse:
      type: object
      properties:
        id:
          type: string
          description: Newsletter JID
        name:
          type: string
          description: Newsletter name
        description:
          type: string
          description: Newsletter description
        subscriberCount:
          type: integer
          description: Number of subscribers
        verificationState:
          type: string
          description: Verification state of the newsletter
        pictureUrl:
          type: string
          description: Profile picture URL
        inviteCode:
          type: string
          description: Invite code for the newsletter
        role:
          type: string
          description: Current user's role in the newsletter
        mute:
          type: string
          description: Mute state for the current user
        state:
          type: string
          description: Newsletter state
        createdAt:
          type: string
          format: date-time
          description: Creation timestamp

    # ─── Status Schemas ─────────────────────────────────────────────
...